Added Wrapper Functions for get_field and get_sub_field
This abstraction layer on these key functions helps keep our underpin
modules easy to update in the future. In this Gutenberg refactor, we
were forced to update many get_sub_field calls to get_field. By adding
this abstraction layer, we are able to add better backwards
compatibility with older versions of Underpin.